+++ src/components/mod.rs
@@ -1,10 +1,12 @@
 mod icon;
+mod icon_button;
 mod state_layer;
 mod surface;
 mod typography;
 
 pub use self::{
     icon::{Icon, IconData, IconStyle},
+    icon_button::{IconButton, IconButtonStyle},
     state_layer::StateLayer,
     surface::Surface,
     typography::Typography,